rewrote interface with the cmd module
